Refactored Auth.tsx, login and register routes created
